requirements.txt
absl-py==2.3.0
accel_hydra
accelerate==1.2.1
alias-free-torch==0.0.6
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens
astunparse==1.6.3
attrs==22.2.0
audioread==3.0.1
av==11.0.0
backcall
beautifulsoup4
bitarray==3.7.1
boto3==1.38.36
botocore==1.38.36
braceexpand==0.1.7
brotlipy==0.7.0
certifi
cffi
chardet
charset-normalizer
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
conda==23.1.0
conda-build==3.23.3
conda-content-trust
conda-package-handling
conda_package_streaming
contourpy==1.2.0
cryptography
cycler==0.12.1
Cython==0.29.37
dcase-util==0.2.20
decorator
diffusers==0.33.1
dnspython==2.3.0
docker-pycreds==0.4.0
# Editable install with no version control (dog==0.1.0)
-e /workspace/work/mypkg
einops==0.7.0
exceptiongroup==1.1.1
executing
expecttest==0.1.4
filelock
fire==0.7.0
flatbuffers==25.12.19
flit_core
fonttools==4.47.2
fsspec==2023.12.2
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
glob2
gmpy2
grpcio==1.73.0
h5py==3.10.0
hdbscan==0.8.37
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
hypothesis==6.70.0
idna
imageio==2.37.0
importlib_metadata==8.5.0
iniconfig==2.0.0
ipdb==0.13.13
ipython
jedi
Jinja2
jmespath==1.0.1
joblib==1.3.2
json5==0.13.0
jsonlines==4.0.0
kaldiio==2.18.1
kiwisolver==1.4.5
laion_clap==1.1.7
lazy-dataset==0.0.14
lazy_loader==0.4
libarchive-c
librosa==0.10.2
llvmlite==0.42.0
lxml==6.0.1
Markdown==3.8
markdown-it-py==4.0.0
MarkupSafe
matplotlib==3.8.2
matplotlib-inline
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.0
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.23.2
packaging==23.2
pandas==2.2.0
parso
pathlib==1.0.1
pexpect
pickleshare
pillow==11.3.0
pkginfo
platformdirs==4.2.1
pluggy==1.5.0
pooch==1.8.1
portalocker==3.2.0
prettytable==3.16.0
progressbar==2.5
prompt-toolkit
protobuf==5.29.2
psds-eval==0.5.3
psutil
ptyprocess
pure-eval
pycosat
pycparser
pydantic==2.10.4
pydantic_core==2.27.2
pydot-ng==2.0.0
pyecharts==2.0.8
Pygments==2.19.2
pynndescent==0.6.0
pynvml==12.0.0
pyOpenSSL
pyparsing==3.1.1
pypinyin==0.55.0
PySocks
pytest==8.2.0
python-dateutil==2.8.2
python-etcd==0.4.5
python-magic==0.4.27
pytz
pyworld==0.3.5
PyYAML
regex==2023.12.25
requests
resampy==0.4.3
rich==13.9.4
ruamel.yaml
ruamel.yaml.clib
s3transfer==0.13.0
sacrebleu==2.5.1
safetensors==0.5.0
scikit-image==0.25.2
scikit-learn==1.4.0
scipy==1.12.0
sed-eval==0.2.1
sed-scores-eval==0.0.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
silero-vad==6.2.0
simplejson==3.20.1
six
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve
soxr==0.3.7
ssr-eval
stack-data
swankit==0.2.3
swanlab==0.6.12
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.2.0
tifffile==2025.5.10
timm==0.9.12
tokenizers==0.21.1
toml
tomli==2.0.1
toolz
torch==2.6.0
torchaudio==2.6.0
torchdata==0.10.1
torchelastic==0.2.2
torchlibrosa==0.1.0
torchtext==0.15.0
torchvision==0.21.0
tqdm
traitlets
transformers==4.51.3
triton==3.2.0
types-dataclasses==0.6.6
typing_extensions==4.12.2
tzdata==2023.4
umap-learn==0.5.6
-e git+https://gitee.com/assasinatee/sss@ab17dec2f9f3a49b1fceb60210b8f68da24ebd3c#egg=uniflow_audio
urllib3
validators==0.28.1
wandb==0.19.1
wcwidth
webdataset==1.0.2
Werkzeug==3.1.3
wespeaker @ git+https://github.com/wenet-e2e/wespeaker.git@d6fd1cbfb51161a558282f3a0effc993c3eaca0d
wget==3.2
wrapt==1.17.2
zipp==3.21.0
zstandard
